backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
openai>=1.3.0
requests>=2.31.0
python-multipart>=0.0.6
numpy>=1.24.0
python-dotenv>=1.0.0
supabase>=2.3.0
websockets>=12.0
aiosqlite>=0.19.0
bcrypt>=4.0.0
pyjwt>=2.8.0
